requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
google-genai>=1.0.0
youtube-transcript-api==0.6.2
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
